MNT: Move code between plan_tools and plans.
